collections/ansible_collections/cisco/dnac/plugins/modules/devicedata.py
import os
import json
import logging
from datetime import datetime
from ansible.module_utils.basic import AnsibleModule
import requests

# Configure logging
# log_dir = os.path.dirname(os.path.abspath(__file__))  # Use the current directory
log_dir = '/home/kethadileep/project_root/cisco_project/logs'
log_file = os.path.join(log_dir, f'devicedata_{datetime.now().strftime("%Y%m%d")}.log')
error_log_file = os.path.join(log_dir, f'devicedata_error_{datetime.now().strftime("%Y%m%d")}.log')

# Set up logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler(log_file),
        logging.FileHandler(error_log_file),
    ]
)

class DeviceDataProcessor:
    def __init__(self, host, username, password, input_data):
        self.host = host
        self.username = username
        self.password = password
        self.input_data = input_data
        self.token = None

    def login(self):
        """Login to the 3rd party API to obtain the token."""
        url = f"https://{self.host}/api/system/v1/auth/token"
        try:
            response = requests.post(url, auth=(self.username, self.password), verify=False)
            response.raise_for_status()
            self.token = response.json().get('Token')
            logging.info("Successfully obtained authentication token.")
        except requests.RequestException as e:
            logging.error(f"Error logging in: {str(e)}")
            raise

    def get_device_data(self):
        """Fetch device data from the API."""
        url = f"https://{self.host}/dna/intent/api/v1/network-device"
        headers = {'Authorization': f"Bearer {self.token}"}
        try:
            response = requests.get(url, headers=headers, verify=False)
            response.raise_for_status()
            logging.info("Successfully fetched device data.")
            return response.json()
        except requests.RequestException as e:
            logging.error(f"Error fetching device data: {str(e)}")
            raise

    def process_data(self):
        """Process and filter the device data."""
        # This method would contain your logic for processing the data.
        pass

def main():
    """Main function to execute the module."""
    module_args = dict(
        host=dict(type='str', required=True),
        username=dict(type='str', required=True),
        password=dict(type='str', required=True),
        input_data=dict(type='dict', required=True),
    )

    module = AnsibleModule(argument_spec=module_args)
    host = module.params['host']
    username = module.params['username']
    password = module.params['password']
    input_data = module.params['input_data']

    processor = DeviceDataProcessor(host, username, password, input_data)

    try:
        processor.login()
        device_data = processor.get_device_data()
        processor.process_data()
        module.exit_json(changed=False, result=device_data)
    except Exception as e:
        module.fail_json(msg=str(e))

if __name__ == '__main__':
    main()
